Implemented basic experiment script
The script expects the participant id as parameter (integer). Upon startup, the explanation for the first task (simple condition) is shown. The participant must press 'Space' when the screen changes its color. Then, the explanation for the second task is shown. During the second task, the participant must press the correct number which is shown on the screen.
All relevant data is logged to 'logs.csv' in the root folder.